BackendServer/requirements.txt
absl-py==0.7.0
astor==0.7.1
audioread==2.1.6
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
docopt==0.6.2
entrypoints==0.3
Flask==1.0.2
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
joblib==0.13.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.27.1
Markdown==3.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
mock==2.0.0
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.4
numba==0.42.1
numpy==1.16.1
opencv-python==4.0.0.21
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pipreqs==0.4.9
prometheus-client==0.5.0
prompt-toolkit==2.0.8
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.8.0
python-mnist==0.6
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
resampy==0.2.1
scikit-learn==0.20.2
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sympy==1.3
tensorboard==1.12.2
tensorflow==1.13.0rc1
tensorflow-estimator==1.13.0rc0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
yarg==0.1.9
